# lib/api-zod/src/generated/types/detection.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BoundingBox } from "./boundingBox";

export interface Detection {
  id: number;
  timestamp: Date;
  personCount: number;
  boxes: BoundingBox[];
}