# lib/api-zod/src/generated/types/detectionStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DetectionStatsRecentActivityItem } from "./detectionStatsRecentActivityItem";

export interface DetectionStats {
  totalDetections: number;
  totalPersonsDetected: number;
  averagePersonsPerDetection: number;
  maxPersonsDetected: number;
  recentActivity: DetectionStatsRecentActivityItem[];
}